Fixed tests and refactored.
