+++ package.json
@@ -17,9 +17,11 @@
         "prepare": "husky install"
     },
     "dependencies": {
+        "@companieshouse/api-sdk-node": "^2.0.70",
         "@companieshouse/node-session-handler": "^4.1.6",
         "@companieshouse/structured-logging-node": "^1.0.4",
         "@companieshouse/web-security-node": "^2.0.0",
+        "axios": "^0.21.4",
         "cookie-parser": "^1.4.6",
         "express": "^4.17.1",
         "govuk-frontend": "^3.14.0",
@@ -27,7 +29,7 @@
         "luxon": "^2.1.1",
         "multer": "^1.4.5-lts.1",
         "nunjucks": "^3.2.3",
-        "private-api-sdk-node": "github:companieshouse/private-api-sdk-node#0.1.71",
+        "private-api-sdk-node": "github:companieshouse/private-api-sdk-node#0.1.76",
         "reflect-metadata": "^0.1.13",
         "uuid": "^8.3.2",
         "yargs": "^15.4.1"
